requirements.txt
fastapi
uvicorn
streamlit
requests
pyyaml
passlib[bcrypt]
pyjwt
python-multipart
bcrypt==3.2.2
passlib